Migrate to electron forge project
