--- README.md
@@ -1,2 +1,0 @@
-# SpeedrunTimerClient
- A highly customizable speedrun timer app. Created as a project for a dev blog series as seen [here.](https://dev.to/kevthedev/making-a-speedrun-timer-prologue-4di)
