+++ index.html
@@ -1,17 +1,12 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="favicon" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Speedrun Timer</title>
-    <meta name="description" content="A highly customizable speedrun timer">
-    <link rel="apple-touch-icon" href="/apple-touch-icon.png" sizes="180x180">
-    <link rel="mask-icon" href="favicon" color="#FFFFFF">
-    <meta name="theme-color" content="#ffffff">
+    <title>Hello World!</title>
+
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="/src/renderer/renderer.js"></script>
   </body>
 </html>
