+++ package.json
@@ -1,24 +1,38 @@
 {
-  "name": "speedruntimerclient",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
+  "name": "my-new-app",
+  "productName": "my-new-app",
+  "version": "1.0.0",
+  "description": "My Electron application description",
+  "main": ".vite/build/main.js",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make",
+    "publish": "electron-forge publish",
+    "lint": "echo \"No linting configured\"",
     "test": "vitest",
     "coverage": "vitest run --coverage"
   },
-  "dependencies": {
-    "vue": "^3.2.47"
+  "keywords": [],
+  "author": {
+    "name": "Kev",
+    "email": "kdevcse@protonmail.com"
   },
+  "license": "MIT",
   "devDependencies": {
+    "@electron-forge/cli": "^6.1.1",
+    "@electron-forge/maker-deb": "^6.1.1",
+    "@electron-forge/maker-rpm": "^6.1.1",
+    "@electron-forge/maker-squirrel": "^6.1.1",
+    "@electron-forge/maker-zip": "^6.1.1",
+    "@electron-forge/plugin-vite": "^6.1.1",
     "@vitejs/plugin-vue": "^4.1.0",
-    "@vitest/coverage-c8": "^0.29.7",
-    "vite": "^4.2.0",
-    "vite-plugin-mkcert": "^1.13.4",
-    "vite-plugin-pwa": "^0.14.6",
-    "vitest": "^0.29.7"
+    "@vitest/coverage-c8": "^0.29.8",
+    "electron": "24.0.0",
+    "vitest": "^0.29.8"
+  },
+  "dependencies": {
+    "electron-squirrel-startup": "^1.0.0",
+    "vue": "^3.2.47"
   }
 }
