+++ vite.preload.config.mjs
@@ -1,0 +1,4 @@
+import { defineConfig } from 'vite';
+
+// https://vitejs.dev/config
+export default defineConfig({});
